pyproject.toml
[tool.poetry]
name = "rss2channel"
version = "0.1.0"
description = ""
authors = ["hootrix <hhtjim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
diskcache = "^5.2.1"
feedparser = "^6.0.8"
pyTelegramBotAPI = "3.6.6"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
